+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Updated codebase to Dart 2.2.
 - `PostgreSQLResult` and `PostgreSQLResultRow` as the return value of a query.
   - Returned lists are protected with `UnmodifiableListView`.
+  - Exposing column metadata through `ColumnDescription`.
 - `PostgreSQLConnection` and `_TransactionProxy` share the OID cache.
 - default value for `query(allowReuse = true)` is set only in the implementation method.